Split up datatore.py into three files
There is no functional change from this. Just code movement
