+++ calico_containers/pycalico/ipam.py
@@ -14,7 +14,8 @@
 
 from etcd import EtcdKeyNotFound, EtcdAlreadyExist
 from netaddr import IPAddress, IPNetwork
-from datastore import DatastoreClient, IPPool
+from datastore import DatastoreClient
+from calico_containers.pycalico.datastore_data import IPPool
 from datastore import CALICO_V_PATH
 
 IP_ASSIGNMENT_PATH = CALICO_V_PATH + "/ipam/%(version)s/assignment/%(pool)s"
