+++ calico_containers/pycalico/netns.py
@@ -22,7 +22,8 @@ import sys
 import uuid
 
 from netaddr import IPNetwork, IPAddress
-from datastore import Endpoint, IF_PREFIX, VETH_NAME
+from datastore import IF_PREFIX, VETH_NAME
+from calico_containers.pycalico.datastore_data import Endpoint
 
 _log = logging.getLogger(__name__)
