+++ calico_containers/tests/unit/datastore_test.py
@@ -26,15 +26,18 @@ from calico_containers.pycalico.datastore import (BGPPeer)
 from calico_containers.pycalico.datastore import (DatastoreClient,
                                                   Rule,
                                                   Profile,
-                                                  Rules,
                                                   Endpoint,
                                                   IPPool,
-                                                  NoEndpointForContainer,
                                                   CALICO_V_PATH,
                                                   DataStoreError,
                                                   ProfileAlreadyInEndpoint,
                                                   ProfileNotInEndpoint,
                                                   MultipleEndpointsMatch)
+from calico_containers.pycalico.datastore_errors import NoEndpointForContainer, \
+    DataStoreError, ProfileNotInEndpoint, ProfileAlreadyInEndpoint, \
+    MultipleEndpointsMatch
+from calico_containers.pycalico.datastore_data import Rules, BGPPeer, IPPool, \
+    Endpoint, Profile, Rule
 
 TEST_HOST = "TEST_HOST"
 TEST_ORCH_ID = "docker"
